rename playSound to useSound
