+++ src/modules.tsx
@@ -4,7 +4,7 @@ import { format, getModules, pretty, round, sumModules } from './calc'
 import { Layout, Square } from './layout'
 import { List, Ul } from './list'
 import Main from './main'
-import playSound from './playSound'
+import useSound from './useSound'
 import { Asset, Chunk, Module } from './stats'
 import { Subtitle, Title } from './typography'
 
@@ -53,7 +53,7 @@ const Modules = ({ chunks, asset }: ModulesProps) => {
   }, [modules])
 
   const handleClick = (isBig: boolean, index: number) => {
-    playSound()
+    useSound()
     setShowBig(isBig)
     setBigModulesIndex(index)
   }
